requirements.txt
Flask>=3.0,<3.1
SpeechRecognition>=3.10.1,<3.11
python-dotenv>=1.0,<1.1
psycopg2>=2.9.9,<2.10